restructure api dir
